Add content of hashed repo to the deps repo
Currently both rdo-trunk-deps-start and rdo-trunk-deps-end contain
only the list of rpms in deps repo. This makes difficult to get
the whole list of RDO packages used for the installation as some
packages may be in both hashed and deps repo and we need to apply
the logic to build the entire list.

This patch extract the entire list of available packages in both
hash and deps repo whith the priorities plugin applied.

Change-Id: I1c282bd16ed5e7223c532696481e825fbcf32f1f