Update patch set 14
Patch Set 14:

I'm not sure if i understand this correctly. My guess is that this role is executed in the host where packstack or o-p-i is going to be run, so in case this is a VM (the case where nested is required), it'd be run inside the vm, am i right?. Note that the "nested" module parameter must be enabled in the underlying physical hypervisor, it's not required in the vm where we want to execute kvm-nested vms.

In fact, taking a look to previous jobs, i'd say that nested virtualization is already enabled in hypervisors in review.r.o:

http://46.231.132.68:8080/v1/AUTH_b50e80d3969f441a8b7b1fe831003e0a/rdoartifacts/01/3201/1/check/gate-weirdo-integration-master-packstack-scenario001/Zf217b0e0b67e4eb8b8fc309f41fb0ba4/artifacts/cpuinfo.txt.gz

So kvm can be configured as virt_type without any of these changes.

Patch-set: 14