Update patch set 14
Patch Set 14:

@Alfredo, the common role always runs before the packstack or puppet-openstack role. It's where the common configuration is done like setting up the trunk or rpm repositories.

This patch will pre-install qemu-kvm and, if it detects to be running in a nested virtualization capable environment (Intel+vmx or AMD+svm), it will reload the kvm kernel module with the nested flag in order to pass the vmx or svm flags down to created virtual machines.

This will work on Baremetal (i.e. ci.centos.org) but also on virtual machines where the hypervisor is also passing the nested config and the appropriate cpu flags.

In our context, tempest is only creating one layer of virtual machines (cirros) and those cirros VMs do not create another VM underneath so we are passing down the actual flags unnecessarily, but this will come in handy for upcoming TripleO cases.
This is mostly about using KVM as the hypervisor instead of qemu in the case of Packstack and Puppet OpenStack.

Patch-set: 14