Fix the repoquery command
After adding delorean-<release>-pending to delorean-deps.repo
disabled, we need to fix repoquery command to avoid getting
this new repo content.

This also fixes the logging location to eb the same as the
other repoquery command so both files are at the same level
of the hierarchy for easy comparison.

Change-Id: Ie4f9404b28f5f7609f934584b372f64da6586959