Rename trunk_repositories fact to trunk_repositories_list
https://review.rdoproject.org/r/#/c/6634/ introduced a new
fact trunk_repositories which contains the list of URLs of
.repo files to configure depending of the content of some
variables including trunk_repositories.

However, this is not working fine when trunk_repositories is
specified via extra-var (-e option) as the variables precedence
for extra-vars is higher that for vars registered using set_fact [1]

This is making some jobs which include -e trunk_repositories=...
to fail as in [2].

This patch renames the fact to avoid any name collision with the
variable added as extra-vars.

[1] http://docs.ansible.com/ansible/playbooks_variables.html#variable-precedence-where-should-i-put-a-variable
[2] https://ci.centos.org/view/rdo/view/all/job/weirdo-generic-packstack-scenario001/611/console

Change-Id: I3400a46d4258f219c9f5e58af2046a4235ddeb2d