Enable EPEL repository for Ceph
Until now, we've enabled EPEL repo for Ceph in puppet-openstack
weirdo role. Now that we also need it enabled to build containers
with tcib in new job [1], I think we should move it to common role
which is the right place to setup repos.
I put it in stable.yml which is run for -release jobs only. If the repo
epel-low-priority does not exist (i.e releases using Ceph prior Bobcat)
the task won't fail, so it's backward compatible.

[1] https://review.rdoproject.org/r/q/topic:okd_container_build

Change-Id: Ia537ae0cfd4523a60d43d5fffc243042d6dd8165
Depends-On: https://review.rdoproject.org/r/c/rdo-jobs/+/51132